Bugfix: Using the new string.format() method
Running the script results in an error from pg8000 like below 

pg8000.core.InterfaceError: '%'' not supported in a quoted string within the query string

This is because pg8000 seems to do some sort of % based replacement again, and that seems to cause an issue. Therefore switching to newer format method so that the %% is retained and when pg8000 encounters the % it doesn't complain. 

Additionally, the old "%" based formatting is on the path to deprecation - http://docs.python.org/tutorial/inputoutput.html#fancier-output-formatting